requirements.txt
selenium
pandas
google-generativeai
openpyxl
webdriver-manager
unidecode
python-dotenv